+++ src/App.tsx
@@ -1,29 +1,8 @@
 import MainLayout from './layouts/MainLayout';
-import AccountPage from './pages/AccountPage';
-import { ErrorPage } from './pages/ErrorPage';
-import { LandingPage } from './pages/LandingPage';
-import { RegisterPage } from './pages/RegisterPage';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import { RouterProvider } from 'react-router-dom';
+import router from './router';
 
 function App() {
-  const router = createBrowserRouter([
-    {
-      path: '/',
-      element: <LandingPage />,
-      errorElement: <ErrorPage />,
-    },
-    {
-      path: '/register',
-      element: <RegisterPage />,
-      errorElement: <ErrorPage />,
-    },
-    {
-      path: '/account',
-      element: <AccountPage />,
-      errorElement: <AccountPage />,
-    },
-  ]);
-
   return (
     <MainLayout>
       <RouterProvider router={router} />
