+++ src/pages/ErrorPage.tsx
@@ -4,12 +4,9 @@ import { Link } from 'react-router-dom';
 export const ErrorPage = () => {
   return (
     <div>
-      <p className="text-bold text-2xl">This is Error Page</p>
+      <p className="text-bold text-2xl text-red-500">This is Error Page</p>
       <Button
         variant="secondary"
-        onClick={() => {
-          alert('Error button clicked... Go to Landing Page');
-        }}
         asChild
         className="bg-black text-white hover:bg-gray-800 transition duration-300"
       >
