+++ src/router.tsx
@@ -1,0 +1,27 @@
+// Router Page
+
+import { createBrowserRouter } from 'react-router-dom';
+import AccountPage from './pages/AccountPage';
+import { ErrorPage } from './pages/ErrorPage';
+import { LandingPage } from './pages/LandingPage';
+import { RegisterPage } from './pages/RegisterPage';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <LandingPage />,
+    errorElement: <ErrorPage />,
+  },
+  {
+    path: '/register',
+    element: <RegisterPage />,
+    errorElement: <ErrorPage />,
+  },
+  {
+    path: '/account',
+    element: <AccountPage />,
+    errorElement: <ErrorPage />,
+  },
+]);
+
+export default router;
